refactor: migrate schedule planning live view
